pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "banner-maker"
version = "0.1.10"
description = "Professional banner generation tool with modern CLI interface"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Onur Barlik"}
]
keywords = ["banner", "generator", "cli", "design", "graphics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "Pillow>=8.0.0",
    "numpy>=1.20.0",
]

[project.urls]
Homepage = "https://github.com/obarlik/banner-maker"
Repository = "https://github.com/obarlik/banner-maker"
Issues = "https://github.com/obarlik/banner-maker/issues"
Documentation = "https://github.com/obarlik/banner-maker/blob/main/README.md"

[project.scripts]
banner-maker = "banner_maker:cli_main"

[tool.setuptools]
packages = ["banner", "core", "demo"]
py-modules = ["banner_maker"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["*.json", "*.ttf", "*.txt"]

